schemaorg4j-parser-core/src/main/java/org/creativemines/schemaorg4j/parser/core/mappers/TimeMapper.java
package org.creativemines.schemaorg4j.parser.core.mappers;

import java.time.LocalTime;
import java.util.Objects;

import org.creativemines.schemaorg4j.parser.core.domain.DataValue;
import org.creativemines.schemaorg4j.parser.core.domain.DataValueString;

public class TimeMapper extends SpecificTypeSchemaMapper {

    @Override
    public boolean canHandle(Class<?> clazz, DataValue argument) {
        return Objects.equals(LocalTime.class, clazz) && argument instanceof DataValueString;
    }

    @Override
    public Object handle(MappingArguments mappingArguments) {
        return ((DataValueString) mappingArguments.getArgument()).asTime();
    }
}
